requirements.txt
# API client libraries
requests>=2.31.0
notion-client>=2.2.1

# Configuration management
pyyaml>=6.0.1

# Environment variable management
python-dotenv>=1.0.0
